+++ semcore/core/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [2.11.0] - 2023-11-29
+
+### Changed
+
+- Renamed type `RemoveFields` to `EfficientOmit` and moved it into `InternalTypings`.
+
 ## [2.10.0] - 2023-11-24
 
 ### Changed