+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.15.0] - 2023-11-29
+
+### Changed
+
+- Deprecated some values of `interaction` property.
+
 ## [4.14.0] - 2023-11-28
 
 ### Added