+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.11.0] - 2023-11-29
+
+### Changed
+
+- Deprecated `interaction` property.
+
 ## [6.10.4] - 2023-11-24
 
 ### Fixed